backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
ffmpeg-python
slowapi
torch
numpy
